refactor: remove --reuseDevice option (#182)
BREAKING CHANGES:

There will be no longer --reuseDevice option available. From now on will be preserved options from appium caps "fullReset"

Migration steps:
To reuse device set in your Appium capabilities file <b>"fullReset": false</b>
Not to reuse device set <b>"fullReset": true</b>

+++ index.ts
@@ -83,8 +83,9 @@ export async function stopServer() {
         await appiumServer.stop();
     }
 
-    if (nsCapabilities.cleanApp) {
+    if (nsCapabilities.cleanApp && !nsCapabilities.ignoreDeviceController) {
         await DeviceController.uninstallApp(nsCapabilities.device, nsCapabilities.appPath);
+        logInfo("Application from device is uninstalled.")
     }
 };
